Refactor integration tests to individual files
